@@ -1,16 +1,16 @@
 module github.com/bwmarrin/discordgo/examples/voice_receive
 
-go 1.22.4
+go 1.24.0
 
 require (
-	github.com/bwmarrin/discordgo v0.28.1
-	github.com/pion/rtp v1.6.0
-	github.com/pion/webrtc/v3 v3.0.0-20200721060053-ca3cc9d940bc
+	github.com/bwmarrin/discordgo v0.29.0
+	github.com/pion/rtp v1.8.22
+	github.com/pion/webrtc/v3 v3.3.6
 )
 
 require (
-	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/gorilla/websocket v1.5.3 // indirect
 	github.com/pion/randutil v0.1.0 // indirect
-	golang.org/x/crypto v0.0.0-20210421170649-83a5a9bb288b // indirect
-	golang.org/x/sys v0.0.0-20201119102817-f84b799fce68 // indirect
+	golang.org/x/crypto v0.42.0 // indirect
+	golang.org/x/sys v0.36.0 // indirect
 )
