@@ -14,55 +14,55 @@ import (
 	"text/template"
 )
 
-var eventHandlerTmpl = template.Must(template.New("eventHandler").Funcs(template.FuncMap{
+var HandlerTmpl = template.Must(template.New("Handler").Funcs(template.FuncMap{
 	"constName":      constName,
 	"isDiscordEvent": isDiscordEvent,
 	"privateName":    privateName,
-}).Parse(`// Code generated by tools/cmd/eventhandlers; DO NOT EDIT
+}).Parse(`// Code generated by tools/cmd/Handlers; DO NOT EDIT
 // See events.go
 
-package gokord
+package event 
 
 // Following are all the event types.
 // Event type values are used to match the events returned by Discord.
 // EventTypes surrounded by __ are synthetic and are internal to DiscordGo.
 const ({{range .}}
-  {{privateName .}}EventType = "{{constName .}}"{{end}}
+  {{.}}Type = "{{constName .}}"{{end}}
 )
 {{range .}}
-// {{privateName .}}EventHandler is an event handler for {{.}} events.
-type {{privateName .}}EventHandler func(*Session, *{{.}})
+// {{privateName .}}Handler is an event handler for {{.}} events.
+type {{privateName .}}Handler func(Session, *{{.}})
 
 // Type returns the event type for {{.}} events.
-func (eh {{privateName .}}EventHandler) Type() string {
-  return {{privateName .}}EventType
+func (eh {{privateName .}}Handler) Type() string {
+  return {{.}}Type
 }
 {{if isDiscordEvent .}}
 // New returns a new instance of {{.}}.
-func (eh {{privateName .}}EventHandler) New() interface{} {
+func (eh {{privateName .}}Handler) New() any {
   return &{{.}}{}
 }{{end}}
 // Handle is the handler for {{.}} events.
-func (eh {{privateName .}}EventHandler) Handle(s *Session, i interface{}) {
+func (eh {{privateName .}}Handler) Handle(s Session, i any) {
   if t, ok := i.(*{{.}}); ok {
     eh(s, t)
   }
 }
 
 {{end}}
-func handlerForInterface(handler interface{}) EventHandler {
+func handlerForInterface(handler FunctionHandler) Handler {
   switch v := handler.(type) {
-  case func(*Session, interface{}):
-    return interfaceEventHandler(v){{range .}}
-  case func(*Session, *{{.}}):
-    return {{privateName .}}EventHandler(v){{end}}
+  case func(Session, any):
+    return interfaceHandler(v){{range .}}
+  case func(Session, *{{.}}):
+    return {{privateName .}}Handler(v){{end}}
   }
 
   return nil
 }
 
 func init() { {{range .}}{{if isDiscordEvent .}}
-  registerInterfaceProvider({{privateName .}}EventHandler(nil)){{end}}{{end}}
+  registerInterfaceProvider({{privateName .}}Handler(nil)){{end}}{{end}}
 }
 `))
 
@@ -82,7 +82,10 @@ func main() {
 		names = append(names, object)
 	}
 	sort.Strings(names)
-	eventHandlerTmpl.Execute(&buf, names)
+	err = HandlerTmpl.Execute(&buf, names)
+	if err != nil {
+		panic(err)
+	}
 
 	src, err := format.Source(buf.Bytes())
 	if err != nil {
@@ -90,7 +93,7 @@ func main() {
 		src = buf.Bytes()
 	}
 
-	err = os.WriteFile(filepath.Join(dir, strings.ToLower("eventhandlers.go")), src, 0644)
+	err = os.WriteFile(filepath.Join(dir, strings.ToLower("event_handlers.go")), src, 0644)
 	if err != nil {
 		log.Fatal(buf, "writing output: %s", err)
 	}
@@ -103,12 +106,11 @@ func constCase(name string) string {
 }
 
 func isDiscordEvent(name string) bool {
-	switch {
-	case name == "Connect", name == "Disconnect", name == "Event", name == "RateLimit", name == "Interface":
-		return false
-	default:
-		return true
-	}
+	return !(name == "Connect" ||
+		name == "Disconnect" ||
+		name == "Event" ||
+		name == "RateLimit" ||
+		name == "Interface")
 }
 
 func constName(name string) string {
