@@ -7,6 +7,7 @@ import (
 	"github.com/nyttikord/gokord/channel"
 	"github.com/nyttikord/gokord/component"
 	"github.com/nyttikord/gokord/discord/types"
+	"github.com/nyttikord/gokord/event"
 	"github.com/nyttikord/gokord/interaction"
 
 	"log"
@@ -43,8 +44,8 @@ var (
 			Description: "Take a survey about modals",
 		},
 	}
-	commandsHandlers = map[string]func(s *gokord.Session, i *gokord.InteractionCreate){
-		"modals-survey": func(s *gokord.Session, i *gokord.InteractionCreate) {
+	commandsHandlers = map[string]func(s event.Session, i *event.InteractionCreate){
+		"modals-survey": func(s event.Session, i *event.InteractionCreate) {
 			err := s.InteractionAPI().Respond(i.Interaction, &interaction.Response{
 				Type: types.InteractionResponseModal,
 				Data: &interaction.ResponseData{
@@ -83,11 +84,11 @@ var (
 )
 
 func main() {
-	s.AddHandler(func(s *gokord.Session, r *gokord.Ready) {
+	s.EventManager().AddHandler(func(s event.Session, r *event.Ready) {
 		log.Println("Bot is up!")
 	})
 
-	s.AddHandler(func(s *gokord.Session, i *gokord.InteractionCreate) {
+	s.EventManager().AddHandler(func(s event.Session, i *event.InteractionCreate) {
 		switch i.Type {
 		case types.InteractionApplicationCommand:
 			if h, ok := commandsHandlers[i.CommandData().Name]; ok {
