@@ -8,6 +8,7 @@ import (
 	"github.com/nyttikord/gokord/channel"
 	"github.com/nyttikord/gokord/discord"
 	"github.com/nyttikord/gokord/discord/types"
+	"github.com/nyttikord/gokord/event"
 	"github.com/nyttikord/gokord/interaction"
 
 	"log"
@@ -222,8 +223,8 @@ var (
 		},
 	}
 
-	commandHandlers = map[string]func(s *gokord.Session, i *gokord.InteractionCreate){
-		"basic-command": func(s *gokord.Session, i *gokord.InteractionCreate) {
+	commandHandlers = map[string]func(s event.Session, i *event.InteractionCreate){
+		"basic-command": func(s event.Session, i *event.InteractionCreate) {
 			s.InteractionAPI().Respond(i.Interaction, &interaction.Response{
 				Type: types.InteractionResponseChannelMessageWithSource,
 				Data: &interaction.ResponseData{
@@ -231,7 +232,7 @@ var (
 				},
 			})
 		},
-		"basic-command-with-files": func(s *gokord.Session, i *gokord.InteractionCreate) {
+		"basic-command-with-files": func(s event.Session, i *event.InteractionCreate) {
 			s.InteractionAPI().Respond(i.Interaction, &interaction.Response{
 				Type: types.InteractionResponseChannelMessageWithSource,
 				Data: &interaction.ResponseData{
@@ -246,7 +247,7 @@ var (
 				},
 			})
 		},
-		"localized-command": func(s *gokord.Session, i *gokord.InteractionCreate) {
+		"localized-command": func(s event.Session, i *event.InteractionCreate) {
 			responses := map[discord.Locale]string{
 				discord.LocaleChineseCN: "你好！ 这是一个本地化的命令",
 			}
@@ -264,7 +265,7 @@ var (
 				panic(err)
 			}
 		},
-		"options": func(s *gokord.Session, i *gokord.InteractionCreate) {
+		"options": func(s event.Session, i *event.InteractionCreate) {
 			// Access options in the order provided by the user.
 			options := i.CommandData().Options
 
@@ -330,8 +331,8 @@ var (
 				},
 			})
 		},
-		"permission-overview": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			perms, err := s.InteractionAPI().CommandPermissions(s.State.User.ID, i.GuildID, i.CommandData().ID)
+		"permission-overview": func(s event.Session, i *event.InteractionCreate) {
+			perms, err := s.InteractionAPI().CommandPermissions(s.SessionState().User().ID, i.GuildID, i.CommandData().ID)
 
 			var restError *gokord.RESTError
 			if errors.As(err, &restError) && restError.Message != nil && restError.Message.Code == discord.ErrCodeUnknownApplicationCommandPermissions {
@@ -408,7 +409,7 @@ var (
 				},
 			})
 		},
-		"subcommands": func(s *gokord.Session, i *gokord.InteractionCreate) {
+		"subcommands": func(s event.Session, i *event.InteractionCreate) {
 			options := i.CommandData().Options
 			content := ""
 
@@ -435,7 +436,7 @@ var (
 				},
 			})
 		},
-		"responses": func(s *gokord.Session, i *gokord.InteractionCreate) {
+		"responses": func(s event.Session, i *event.InteractionCreate) {
 			// Responses to a command are very important.
 			// First of all, because you need to react to the interaction
 			// by sending the response in 3 seconds after receiving, otherwise
@@ -493,7 +494,7 @@ var (
 				s.InteractionAPI().ResponseDelete(i.Interaction)
 			})
 		},
-		"followups": func(s *gokord.Session, i *gokord.InteractionCreate) {
+		"followups": func(s event.Session, i *event.InteractionCreate) {
 			// Followup messages are basically regular messages (you can create as many of them as you wish)
 			// but work as they are created by webhooks and their functionality
 			// is for handling additional messages after sending a response.
@@ -538,7 +539,7 @@ var (
 )
 
 func init() {
-	s.AddHandler(func(s *gokord.Session, i *gokord.InteractionCreate) {
+	s.EventManager().AddHandler(func(s event.Session, i *event.InteractionCreate) {
 		if h, ok := commandHandlers[i.CommandData().Name]; ok {
 			h(s, i)
 		}
@@ -546,8 +547,8 @@ func init() {
 }
 
 func main() {
-	s.AddHandler(func(s *gokord.Session, r *gokord.Ready) {
-		log.Printf("Logged in as: %v#%v", s.State.User.Username, s.State.User.Discriminator)
+	s.EventManager().AddHandler(func(s event.Session, r *event.Ready) {
+		log.Printf("Logged in as: %v#%v", s.SessionState().User().Username, s.SessionState().User().Discriminator)
 	})
 	err := s.Open()
 	if err != nil {
@@ -557,7 +558,7 @@ func main() {
 	log.Println("Adding commands...")
 	registeredCommands := make([]*interaction.Command, len(commands))
 	for i, v := range commands {
-		cmd, err := s.InteractionAPI().CommandCreate(s.State.User.ID, *GuildID, v)
+		cmd, err := s.InteractionAPI().CommandCreate(s.SessionState().User().ID, *GuildID, v)
 		if err != nil {
 			log.Panicf("Cannot create '%v' command: %v", v.Name, err)
 		}
@@ -583,7 +584,7 @@ func main() {
 		// }
 
 		for _, v := range registeredCommands {
-			err := s.InteractionAPI().CommandDelete(s.State.User.ID, *GuildID, v.ID)
+			err := s.InteractionAPI().CommandDelete(s.SessionState().User().ID, *GuildID, v.ID)
 			if err != nil {
 				log.Panicf("Cannot delete '%v' command: %v", v.Name, err)
 			}
